rename all script files due to live recomendations
