+++ src/main.js
@@ -1,12 +1,12 @@
-import MenuView from './view/Menu.js';
-import TripInfoView from './view/TripInfo.js';
-import FiltersView from './view/Filters.js';
-import SortView from './view/Sort.js';
-import TripPointEditorView from './view/TripPointEditor.js';
-import TripPointView from './view/TripPoint.js';
-import TripPointsContainerView from './view/TripPointsContainer.js';
-import TripPointsContainerEmptyView from './view/TripPointsContainerEmpty.js';
-import ViewElementWrapper from './view/ViewElementWrapper.js';
+import MenuView from './view/menu.js';
+import TripInfoView from './view/trip-info.js';
+import FiltersView from './view/filters.js';
+import SortView from './view/sort.js';
+import TripPointEditorView from './view/trip-point-editor.js';
+import TripPointView from './view/trip-point.js';
+import TripPointsContainerView from './view/trip-points-container.js';
+import TripPointsContainerEmptyView from './view/trip-points-container-empty.js';
+import ViewElementWrapper from './view/view-element-wrapper.js';
 import { RenderPosition } from './utils.js';
 import { generateTripPointData } from './mock/trip-point.js';
 import { ViewValues } from './constants.js';
