+++ src/view/trip-point-editor.js
@@ -1,4 +1,4 @@
-import ViewElement from './ViewElement.js';
+import ViewElement from './abstract-view-element.js';
 import { appData } from '../app-data.js';
 import { TimeUtils } from '../utils.js';
