+++ src/view/trip-points-container-empty.js
@@ -1,4 +1,4 @@
-import ViewElement from './ViewElement.js';
+import ViewElement from './abstract-view-element.js';
 
 export default class TripPointsContainerEmpty extends ViewElement {
   constructor() {
